aaudio: fix command thread busy loop.
When a shared stream is idle, it'll be busy retring entering standby.
The reason is that there is no need to enter standby mode for shared
stream. In that case, set standbyTime to max() if standby_l is not
implemented. For mmap stream, the standby_l can return error if there is
something wrong with the HAL. In that case, wait some more time to retry
entering standby for mmap stream.

Test: atest AAudioTests
Test: Play Netflix video and check `top -H`
Bug: 260165213
Change-Id: I041d1b146afdffa0ba62b604a8fe6c0f59e1ddd9
Merged-In: I041d1b146afdffa0ba62b604a8fe6c0f59e1ddd9
Signed-off-by: Tkpointz <pointzztk@gmail.com>